Explictly cast to unsigned int to suit printf format

+++ vmc/iisc.c
@@ -221,7 +221,7 @@ IIScPrint(
 
   printf(
     "\nDump of sc @ 0x%05x, %d entr%s, current max %d\nIndex\tKey\t\tValue\n",
-    sc, sc->count, sc->count == 1 ? "y" : "ies",  sc->maxCount);
+    ((unsigned int)sc), sc->count, sc->count == 1 ? "y" : "ies",  sc->maxCount);
   for (index = 0; index < sc->size; index++) {
     key = sc->table[index].key;
     value = sc->table[index].value;
